Pull out tests into their own file
