+++ Sources/SwiftInspectorVisitors/Tests/ProtocolVisitorSpec.swift
@@ -165,18 +165,9 @@ final class ProtocolVisitorSpec: QuickSpec {
             var associatedTypeNameToInfoMap: [String: AssociatedtypeInfo]?
             beforeEach {
               let content = """
-                public protocol Collection {
-                  typealias IndexDistance = Int
-
-                  associatedtype Element
-
-                  associatedtype Index : Comparable where Self.Index == Self.Indices.Element
-
-                  associatedtype Iterator = IndexingIterator<Self>
-
-                  associatedtype SubSequence : Collection = Slice<Self> where Self.Element == Self.SubSequence.Element, Self.SubSequence == Self.SubSequence.SubSequence
-
-                  associatedtype Indices : Collection = DefaultIndices<Self> where Self.Indices == Self.Indices.SubSequence
+                public protocol Transformer {
+                  associatedtype Input
+                  associatedtype Output
                 }
                 """
 
@@ -190,71 +181,12 @@ final class ProtocolVisitorSpec: QuickSpec {
                 }
             }
 
-            it("finds the associatedtype Element") {
-              expect(associatedTypeNameToInfoMap?["Element"]).toNot(beNil())
+            it("finds the associatedtype Input") {
+              expect(associatedTypeNameToInfoMap?["Input"]).toNot(beNil())
             }
 
-            it("finds the associatedtype Index") {
-              expect(associatedTypeNameToInfoMap?["Index"]).toNot(beNil())
-            }
-
-            it("finds the associatedtype Index's inheritance") {
-              expect(associatedTypeNameToInfoMap?["Index"]?.inheritsFromTypes.count) == 1
-              expect(associatedTypeNameToInfoMap?["Index"]?.inheritsFromTypes.first?.asSource) == "Comparable"
-            }
-
-            it("finds the associatedtype Index's generic requirements") {
-              expect(associatedTypeNameToInfoMap?["Index"]?.genericRequirements.count) == 1
-              expect(associatedTypeNameToInfoMap?["Index"]?.genericRequirements.first?.leftType.asSource) == "Self.Index"
-              expect(associatedTypeNameToInfoMap?["Index"]?.genericRequirements.first?.rightType.asSource) == "Self.Indices.Element"
-            }
-
-            it("finds the associatedtype Iterator") {
-              expect(associatedTypeNameToInfoMap?["Iterator"]).toNot(beNil())
-            }
-
-            it("finds the associatedtype Iterator's initializer") {
-              expect(associatedTypeNameToInfoMap?["Iterator"]?.initializer?.asSource) == "IndexingIterator<Self>"
-            }
-
-            it("finds the associatedtype SubSequence") {
-              expect(associatedTypeNameToInfoMap?["SubSequence"]).toNot(beNil())
-            }
-
-            it("finds the associatedtype SubSequence's inheritance") {
-              expect(associatedTypeNameToInfoMap?["SubSequence"]?.inheritsFromTypes.count) == 1
-              expect(associatedTypeNameToInfoMap?["SubSequence"]?.inheritsFromTypes.first?.asSource) == "Collection"
-            }
-
-            it("finds the associatedtype SubSequence's generic requirements") {
-              expect(associatedTypeNameToInfoMap?["SubSequence"]?.genericRequirements.count) == 2
-              expect(associatedTypeNameToInfoMap?["SubSequence"]?.genericRequirements.first?.leftType.asSource) == "Self.Element"
-              expect(associatedTypeNameToInfoMap?["SubSequence"]?.genericRequirements.first?.relationship) == .equals
-              expect(associatedTypeNameToInfoMap?["SubSequence"]?.genericRequirements.first?.rightType.asSource) == "Self.SubSequence.Element"
-              expect(associatedTypeNameToInfoMap?["SubSequence"]?.genericRequirements.last?.leftType.asSource) == "Self.SubSequence"
-              expect(associatedTypeNameToInfoMap?["SubSequence"]?.genericRequirements.last?.relationship) == .equals
-              expect(associatedTypeNameToInfoMap?["SubSequence"]?.genericRequirements.last?.rightType.asSource) == "Self.SubSequence.SubSequence"
-            }
-
-            it("finds the associatedtype Indices") {
-              expect(associatedTypeNameToInfoMap?["Indices"]).toNot(beNil())
-            }
-
-            it("finds the associatedtype Indices's inheritance") {
-              expect(associatedTypeNameToInfoMap?["Indices"]?.inheritsFromTypes.count) == 1
-              expect(associatedTypeNameToInfoMap?["Indices"]?.inheritsFromTypes.first?.asSource) == "Collection"
-            }
-
-            it("finds the associatedtype Indices's inheritance") {
-              expect(associatedTypeNameToInfoMap?["Indices"]?.inheritsFromTypes.count) == 1
-              expect(associatedTypeNameToInfoMap?["Indices"]?.initializer?.asSource) == "DefaultIndices<Self>"
-            }
-
-            it("finds the associatedtype Indices's generic requirement") {
-              expect(associatedTypeNameToInfoMap?["Indices"]?.genericRequirements.count) == 1
-              expect(associatedTypeNameToInfoMap?["Indices"]?.genericRequirements.first?.leftType.asSource) == "Self.Indices"
-              expect(associatedTypeNameToInfoMap?["Indices"]?.genericRequirements.first?.relationship) == .equals
-              expect(associatedTypeNameToInfoMap?["Indices"]?.genericRequirements.first?.rightType.asSource) == "Self.Indices.SubSequence"
+            it("finds the associatedtype Output") {
+              expect(associatedTypeNameToInfoMap?["Output"]).toNot(beNil())
             }
           }
         }
